added circle functions and last fix of int to doubles

+++ include/data.h
@@ -13,6 +13,6 @@ void	set_start_position_player(t_all *data);
 bool	player_found(char c);
 
 // -- Utils
-void	set_position(int x, int y, int where[2]);
+void	set_position(int x, int y, double where[2]);
 
 #endif
